ai 대화창 모듈화
